[SM][DLS-4640]Fixed tests and cleaned them up.
